+++ lib/feature_screen_arguments/model/screen_arguments.dart
@@ -1,0 +1,8 @@
+// You can pass any object to the arguments parameter. In this example,
+// create a class that contains both a customizable title and message.
+class ScreenArguments {
+  final String title;
+  final String message;
+
+  ScreenArguments(this.title, this.message);
+}
